pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "taalcr"
version = "0.2.0"
description = "Toolkit for Aggregate Analysis of Language in Conversation, for Research"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Nick McCloskey" }]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "numpy==1.26.4",
    "pandas>=2.0.3",
    "pingouin==0.5.5",
    "scikit_learn>=1.7.2",
    "scipy>=1.16.2",
    "spacy==3.8.7",
    "streamlit>=1.45.1",
    "tqdm>=4.67.1",
    "rascal_speech==0.2.0",
    "spacy[transformers]>=3.8"
]

[project.scripts]
taalcr = "taalcr.cli:main"

[project.optional-dependencies]
dev = ["pytest"]